Phase3/restaurant/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
mssql-django==1.5
packaging==24.1
pyodbc==5.1.0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
uritemplate==4.1.1
